pyproject.toml
[project]
name = "docker"
version = "0.1.0"
description = ""
authors = [
    {name = "86188", email = "w18827306471@163.com"}
]
requires-python = ">=3.10,<3.14"
dependencies = [
    "fastapi>=0.135.1,<0.136.0",
    "uvicorn>=0.41.0,<0.42.0",
    "pydantic>=2.12.5,<3.0.0",
    "requests>=2.32.5,<3.0.0",
    "httpx>=0.28.1,<0.29.0",
    "redis-om>=1.0.6,<2.0.0",
    "pymilvus>=2.6.9,<3.0.0",
    "elasticsearch-dsl>=8.18.0,<9.0.0",
    "openai>=2.24.0,<3.0.0",
    "chainlit>=2.9.6,<3.0.0",
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"